Split agent run command tests by OS
