+++ test/fixture/config/agent_run_windows.conf
@@ -1,9 +1,8 @@
 agent {
-  data_dir = "./.data"
+  insecure_bootstrap = true
   log_level = "DEBUG"
   server_address = "127.0.0.1"
   server_port = "8081"
-  socket_path ="/tmp/spire-agent/public/api.sock"
   trust_domain = "example.org"
 }
 
@@ -22,7 +21,7 @@ plugins {
       kubelet_read_only_port = "10255"
     }
   }
-  WorkloadAttestor "unix" {
+  WorkloadAttestor "windows" {
     plugin_data {
     }
   }
